requirements.txt
asgiref==3.2.3
astroid==2.3.3
Babel==2.8.0
Django==3.0.3
django-bootstrap-modal-forms==1.5.0
django-class-based-auth-views==0.4
django-phonenumber-field==4.0.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
phonenumbers==8.12.0
Pillow==7.0.0
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
wrapt==1.11.2
